// types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      assistants: {
        Row: {
          created_at: string
          description: string | null
          id: string
          name: string
          openai_assistant_id: string | null
          openai_vector_store_id: string | null
          system_prompt: string | null
          user_prompt: string | null
          welcome_message: string[] | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: string
          name: string
          openai_assistant_id?: string | null
          openai_vector_store_id?: string | null
          system_prompt?: string | null
          user_prompt?: string | null
          welcome_message?: string[] | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: string
          name?: string
          openai_assistant_id?: string | null
          openai_vector_store_id?: string | null
          system_prompt?: string | null
          user_prompt?: string | null
          welcome_message?: string[] | null
        }
        Relationships: []
      }
// ...
      messages: {
        Row: {
          assistant_id: string | null
          completed: boolean
          content: string | null
          created_at: string
          id: string
          metadata: Json | null
          role: string
          thread_id: string
          user_id: string | null
        }
        Insert: {
          assistant_id?: string | null
          completed?: boolean
          content?: string | null
          created_at?: string
          id?: string
          metadata?: Json | null
          role: string
          thread_id: string
          user_id?: string | null
        }
        Update: {
          assistant_id?: string | null
          completed?: boolean
          content?: string | null
          created_at?: string
          id?: string
          metadata?: Json | null
          role?: string
          thread_id?: string
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "messages_assistant_id_fkey"
            columns: ["assistant_id"]
            isOneToOne: false
            referencedRelation: "assistants"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "messages_thread_id_fkey"
            columns: ["thread_id"]
            isOneToOne: false
            referencedRelation: "threads"
            referencedColumns: ["id"]
          },
        ]
      }
      threads: {
        Row: {
          assistant_id: string | null
          created_at: string
          id: string
          metadata: Json | null
          title: string | null
          updated_at: string
          user_id: string
        }
        Insert: {
          assistant_id?: string | null
          created_at?: string
          id?: string
          metadata?: Json | null
          title?: string | null
          updated_at?: string
          user_id: string
        }
        Update: {
          assistant_id?: string | null
          created_at?: string
          id?: string
          metadata?: Json | null
          title?: string | null
          updated_at?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "threads_assistant_id_fkey"
            columns: ["assistant_id"]
            isOneToOne: false
            referencedRelation: "assistants"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      increment_job_processed_emails: {
        Args: { p_job_id: string; p_increment_amount?: number }
        Returns: undefined
      }
    }
    Enums: {
      email_priority: "LOW" | "MID" | "HIGH"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}